mcp_server/TestScenarioGenerator/requirements.txt
# Core dependencies
langchain-core>=0.3.0
langchain>=0.3.0
python-dotenv>=1.0.0

# Provider-specific dependencies
langchain-openai>=0.3.0
langchain-google-genai>=0.3.0
langchain-anthropic>=0.3.0

# HTTP client for API calls
httpx>=0.24.0 